Factor out interp_utils.*.
