+++ Include/interp_utils.h
@@ -1,0 +1,16 @@
+#ifndef Py_INTERP_UTILS_H
+#define Py_INTERP_UTILS_H
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#ifndef Py_LIMITED_API
+#  define Py_CPYTHON_INTERP_UTILS_H
+#  include "cpython/interp_utils.h"
+#  undef Py_CPYTHON_INTERP_UTILS_H
+#endif
+
+#ifdef __cplusplus
+}
+#endif
+#endif /* !Py_INTERP_UTILS_H */
